+++ Expanded Cheat Menu/HelperMod/BinAssets/AssetsRevision.lua
@@ -20,6 +20,7 @@ CreateRealTimeThread(function()
 
 	-- remove blacklist for any mods in "Mod Ids"
 	local StringFormat = string.format
+	local LuaRevision = LuaRevision
 	for _,mod in pairs(Mods) do
 		if mod_ids[mod.steam_id] then
 			-- i don't set this in mod\metadata.lua so it gives an error